pinba-php deploy

+++ net-misc/pinba/pinba-1.0.0.ebuild
@@ -27,8 +27,8 @@ pkg_setup() {
 	# ugly mysql headers hack
 	einfo "creating symlinks of mysql header files under /usr/include/mysql"
 	for header in my_bitmap.h my_compare.h myisampack.h ft_global.h;do
-	if [[ -e /usr/include/mysql/private/${header} ]] ;then
-		ln -sv /usr/include/mysql/private/${header} /usr/include/mysql/${header}
+	if [[ -f /usr/include/mysql/private/${header} ]] ;then
+		[[ -f /usr/include/mysql/${header} ]] || ln -sv /usr/include/mysql/private/${header} /usr/include/mysql/${header}
 	fi
 	done
 }
